requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
apturl==0.5.2
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
backcall==0.2.0
bcrypt==3.1.7
bidict==0.22.1
blinker==1.7.0
Brlapi==0.7.0
certifi==2019.11.28
chardet==3.0.4
click==8.1.7
colorama==0.4.3
comm==0.2.1
command-not-found==0.3
contourpy==1.1.1
cryptography==2.8
cupshelpers==1.0
cycler==0.12.1
Cython==3.0.8
dbus-python==1.2.16
debugpy==1.8.0
decorator==5.1.1
defer==1.0.6
distro==1.4.0
distro-info===0.23ubuntu1
duplicity==0.8.12.0
entrypoints==0.3
executing==2.0.1
fail2ban==0.11.1
fasteners==0.14.1
filelock==3.13.1
flask==3.0.0
Flask-SocketIO==5.3.6
fonttools==4.47.2
frozenlist==1.4.1
fsspec==2023.12.2
future==0.18.2
gurobipy==11.0.0
h11==0.14.0
httplib2==0.14.0
idna==2.8
importlib-metadata==7.0.1
importlib-resources==6.1.1
ipykernel==6.29.0
ipython==8.12.3
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jupyter-client==8.6.0
jupyter-core==5.7.1
keyring==18.0.1
kiwisolver==1.4.5
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazytensor==0.1.0a2
lightning-utilities==0.10.0
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==2.1.3
matplotlib==3.7.4
matplotlib-inline==0.1.6
monotonic==1.5
mpmath==1.3.0
multidict==6.0.4
nest-asyncio==1.5.9
netifaces==0.10.4
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.1.0
olefile==0.46
opencv-python==4.9.0.80
packaging==23.2
pandas==2.0.3
paramiko==2.6.0
parso==0.8.3
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
platformdirs==4.1.0
POT==0.9.3
prompt-toolkit==3.0.43
protobuf==3.6.1
psutil==5.9.7
pure-eval==0.2.2
pycairo==1.16.2
pycups==1.9.73
pygments==2.17.2
PyGObject==3.36.0
pyinotify==0.9.6
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==3.1.1
pyRFC3339==1.1
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-engineio==4.8.2
python-socketio==5.11.0
pytorch-lightning==2.1.3
pytz==2023.3.post1
pyxdg==0.26
PyYAML==6.0.1
pyzmq==25.1.2
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.1
SecretStorage==2.3.1
simple-websocket==1.0.0
simplejson==3.16.0
six==1.14.0
ssh-import-id==5.10
stack-data==0.6.3
sympy==1.12
systemd-python==234
threadpoolctl==3.2.0
torch==2.1.2
torchmetrics==1.3.0
torchvision==0.16.2
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
triton==2.1.0
typing-extensions==4.9.0
tzdata==2023.4
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
wadllib==1.3.3
wcwidth==0.2.13
werkzeug==3.0.1
wsproto==1.2.0
xkit==0.0.0
yarl==1.9.4
zipp==3.17.0
